Only require password when lid is closed
This means that password is not required when screen turns off. This
behavior can be achieved more easily by changing Windows settings. In
Windows 10 you can do it via:

  Sign-in options -> Require sign-in -> Every Time

So I modified this behavior to only require password when I'm closing
the lid, as I don't like the original behavior. My screen turns off
after a few minutes when on battery, and usually I'm still sitting near
my computer, so typing in password every time is annoying. And when I'm
closing the lid, I want to be sure that computer is locked every time
and instantly (which is not true if you set the above option from "Every
time" to something else like "5 minutes").